app/requirements.txt
pandas
requests
google-cloud-bigquery
google-cloud-bigquery-storage
streamlit
db-dtypes